Getting rid of SummaryComponentSwitch, now that every component type can render itself in a Summary
