+++ src/components/summary/SummaryComponent.tsx
@@ -6,8 +6,10 @@ import cn from 'classnames';
 import { useAppDispatch } from 'src/common/hooks/useAppDispatch';
 import { useAppSelector } from 'src/common/hooks/useAppSelector';
 import { ErrorPaper } from 'src/components/message/ErrorPaper';
-import { SummaryComponentSwitch } from 'src/components/summary/SummaryComponentSwitch';
+import { SummaryBoilerplate } from 'src/components/summary/SummaryBoilerplate';
 import { FormLayoutActions } from 'src/features/form/layout/formLayoutSlice';
+import { GenericComponent } from 'src/layout/GenericComponent';
+import { FormComponent } from 'src/layout/LayoutComponent';
 import { pageBreakStyles } from 'src/utils/formComponentUtils';
 import { useResolvedNode } from 'src/utils/layout/ExprContext';
 import { getTextFromAppOrDefault } from 'src/utils/textResource';
@@ -115,6 +117,9 @@ export function SummaryComponent({ summaryNode, overrides }: ISummaryComponent) 
   const displayGrid =
     display && display.useComponentGrid ? overrides?.grid || targetItem?.grid : overrides?.grid || grid;
 
+  const component = targetNode.getComponent();
+  const RenderSummary = component instanceof FormComponent ? component.renderSummary.bind(component) : null;
+
   return (
     <Grid
       item={true}
@@ -132,13 +137,27 @@ export function SummaryComponent({ summaryNode, overrides }: ISummaryComponent) 
           [classes.border]: !display?.hideBottomBorder,
         })}
       >
-        <SummaryComponentSwitch
-          summaryNode={summaryNode}
-          targetNode={targetNode}
-          change={change}
-          label={label}
-          overrides={overrides}
-        />
+        {RenderSummary && component instanceof FormComponent ? (
+          <>
+            {component.renderSummaryBoilerplate() ? (
+              <SummaryBoilerplate
+                {...change}
+                label={label}
+                summaryNode={summaryNode}
+                targetNode={targetNode}
+                overrides={overrides}
+              />
+            ) : null}
+            <RenderSummary
+              {...change}
+              summaryNode={summaryNode}
+              targetNode={targetNode as any}
+              overrides={overrides}
+            />
+          </>
+        ) : (
+          <GenericComponent node={targetNode} />
+        )}
         {targetNode?.hasValidationMessages('errors') &&
           targetItem.type !== 'Group' &&
           !display?.hideValidationMessages && (
